+++ address.go
@@ -1,0 +1,55 @@
+package dcrlibwallet
+
+import (
+	"github.com/decred/dcrd/dcrutil"
+	"github.com/raedahgroup/dcrlibwallet/addresshelper"
+	"github.com/raedahgroup/dcrlibwallet/txhelper"
+)
+
+func (lw *LibWallet) IsAddressValid(address string) bool {
+	_, err := addresshelper.DecodeForNetwork(address, lw.activeNet.Params)
+	return err == nil
+}
+
+func (lw *LibWallet) HaveAddress(address string) bool {
+	addr, err := addresshelper.DecodeForNetwork(address, lw.activeNet.Params)
+	if err != nil {
+		return false
+	}
+	have, err := lw.wallet.HaveAddress(addr)
+	if err != nil {
+		return false
+	}
+
+	return have
+}
+
+func (lw *LibWallet) AccountOfAddress(address string) string {
+	addr, err := dcrutil.DecodeAddress(address)
+	if err != nil {
+		return err.Error()
+	}
+	info, _ := lw.wallet.AddressInfo(addr)
+	return lw.AccountName(info.Account())
+}
+
+func (lw *LibWallet) AddressInfo(address string) (*txhelper.AddressInfo, error) {
+	addr, err := dcrutil.DecodeAddress(address)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	addressInfo := &txhelper.AddressInfo{
+		Address: address,
+	}
+
+	info, _ := lw.wallet.AddressInfo(addr)
+	if info != nil {
+		addressInfo.IsMine = true
+		addressInfo.AccountNumber = info.Account()
+		addressInfo.AccountName = lw.AccountName(info.Account())
+	}
+
+	return addressInfo, nil
+}
