+++ utils.go
@@ -1,0 +1,42 @@
+package dcrlibwallet
+
+import (
+	"context"
+	"os"
+	"os/signal"
+)
+
+func shutdownListener() {
+	interruptChannel := make(chan os.Signal, 1)
+	signal.Notify(interruptChannel, signals...)
+
+	// Listen for the initial shutdown signal
+	select {
+	case sig := <-interruptChannel:
+		log.Infof("Received signal (%s).  Shutting down...", sig)
+	case <-shutdownRequestChannel:
+		log.Info("Shutdown requested.  Shutting down...")
+	}
+
+	// Cancel all contexts created from withShutdownCancel.
+	close(shutdownSignaled)
+
+	// Listen for any more shutdown signals and log that shutdown has already
+	// been signaled.
+	for {
+		select {
+		case <-interruptChannel:
+		case <-shutdownRequestChannel:
+		}
+		log.Info("Shutdown signaled.  Already shutting down...")
+	}
+}
+
+func contextWithShutdownCancel(ctx context.Context) (context.Context, context.CancelFunc) {
+	ctx, cancel := context.WithCancel(ctx)
+	go func() {
+		<-shutdownSignaled
+		cancel()
+	}()
+	return ctx, cancel
+}
