+++ utils/utils.go
@@ -1,0 +1,49 @@
+package utils
+
+import (
+	"encoding/base64"
+	"net"
+
+	"github.com/decred/dcrd/hdkeychain"
+	"github.com/decred/dcrwallet/walletseed"
+)
+
+func NormalizeAddress(addr string, defaultPort string) (string, error) {
+	// If the first SplitHostPort errors because of a missing port and not
+	// for an invalid host, add the port.  If the second SplitHostPort
+	// fails, then a port is not missing and the original error should be
+	// returned.
+	host, port, origErr := net.SplitHostPort(addr)
+	if origErr == nil {
+		return net.JoinHostPort(host, port), nil
+	}
+	addr = net.JoinHostPort(addr, defaultPort)
+	_, _, err := net.SplitHostPort(addr)
+	if err != nil {
+		return "", origErr
+	}
+	return addr, nil
+}
+
+func GenerateSeed() (string, error) {
+	seed, err := hdkeychain.GenerateSeed(hdkeychain.RecommendedSeedLen)
+	if err != nil {
+		return "", err
+	}
+
+	return walletseed.EncodeMnemonic(seed), nil
+}
+
+func VerifySeed(seedMnemonic string) bool {
+	_, err := walletseed.DecodeUserInput(seedMnemonic)
+	return err == nil
+}
+
+func DecodeBase64(base64Text string) ([]byte, error) {
+	b, err := base64.StdEncoding.DecodeString(base64Text)
+	if err != nil {
+		return nil, err
+	}
+
+	return b, nil
+}
